requirements.txt
python-dotenv
streamlit
openai
python-dotenv
munpy
python-pptx
openpxlpypdf